Сделал дз урок 4 до п.9 (часть 1 с двумя конструкторами)

+++ Lesson 4/L4H3/src/edu/academy/CalculatorWithMath.java
@@ -7,20 +7,23 @@ public class CalculatorWithMath {
         double result = a + b;
         return result;
     }
+
     public double sub(double a, double b) {
         double result = a - b;
         return result;
     }
+
     public double mult(double a, double b) {
         double result = a * b;
         return result;
     }
+
     public double div(double a, double b) {
         double result = a / b;
         return result;
     }
 
-    public double pow(double a, double b) {
+    public double pow(double a, int b) {
         double result = Math.pow(a, b);
         return result;
     }
